fix(PROD-814): 뮤트 목록의 화면 구성 책임을 분리한다
