+++ apps/app/src/components/settings/SettingsMutedProfiles.tsx
@@ -3,12 +3,5 @@ import { MutedProfileList } from '@/components/profile/MutedProfileList';
 const noopUnmute = () => Promise.resolve();
 
 export function SettingsMutedProfiles() {
-  return (
-    <MutedProfileList
-      onUnmute={noopUnmute}
-      scrollable={false}
-      showHeading={false}
-      state={{ status: 'loading' }}
-    />
-  );
+  return <MutedProfileList onUnmute={noopUnmute} state={{ status: 'loading' }} />;
 }
